+++ src/main/java/forestry/arboriculture/ModuleArboriculture.java
@@ -10,12 +10,16 @@
  ******************************************************************************/
 package forestry.arboriculture;
 
+import java.util.Arrays;
 import java.util.function.Consumer;
 
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.DispenserBlock;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.WoodType;
 
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
@@ -23,6 +27,7 @@ import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.common.capabilities.RegisterCapabilitiesEvent;
 import net.minecraftforge.event.AttachCapabilitiesEvent;
+import net.minecraftforge.event.TagsUpdatedEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -37,8 +42,10 @@ import forestry.api.genetics.IIndividual;
 import forestry.api.modules.ForestryModule;
 import forestry.api.modules.ForestryModuleIds;
 import forestry.api.modules.IPacketRegistry;
+import forestry.arboriculture.blocks.BlockDecorativeLeaves;
 import forestry.arboriculture.client.ArboricultureClientHandler;
 import forestry.arboriculture.commands.CommandTree;
+import forestry.arboriculture.features.ArboricultureBlocks;
 import forestry.arboriculture.features.ArboricultureItems;
 import forestry.arboriculture.items.ForestryBoatDispenserBehavior;
 import forestry.arboriculture.network.PacketRipeningUpdate;
@@ -47,6 +54,7 @@ import forestry.core.genetics.capability.IndividualHandlerItem;
 import forestry.core.network.PacketIdClient;
 import forestry.core.utils.SpeciesUtil;
 import forestry.modules.BlankForestryModule;
+import forestry.modules.features.FeatureBlock;
 
 @ForestryModule
 public class ModuleArboriculture extends BlankForestryModule {
@@ -61,6 +69,7 @@ public class ModuleArboriculture extends BlankForestryModule {
 
 		modBus.addListener(ModuleArboriculture::registerCapabilities);
 		modBus.addListener(ModuleArboriculture::commonSetup);
+		MinecraftForge.EVENT_BUS.addListener(ModuleArboriculture::modifyLeafStateCaches);
 		MinecraftForge.EVENT_BUS.addGenericListener(ItemStack.class, ModuleArboriculture::attachCapabilities);
 	}
 
@@ -74,6 +83,19 @@ public class ModuleArboriculture extends BlankForestryModule {
 
 			if (individual != null) {
 				event.addCapability(IIndividual.CAPABILITY_ID, new IndividualHandlerItem(type, stack, individual, TreeLifeStage.SAPLING));
+			}
+		}
+	}
+
+	// Called right after Block.rebuildCache to modify BlockStateBase.Cache, preventing fences from connecting to Forestry decorative leaves
+	private static void modifyLeafStateCaches(TagsUpdatedEvent event) {
+		// Use one copy of cache for all blocks/states to avoid wasting memory
+		BlockBehaviour.BlockStateBase.Cache leafCache = ArboricultureBlocks.LEAVES_DECORATIVE.getList().get(0).defaultBlockState().cache;
+		Arrays.fill(leafCache.faceSturdy, false);
+
+		for (FeatureBlock<BlockDecorativeLeaves, BlockItem> feature : ArboricultureBlocks.LEAVES_DECORATIVE.getFeatures()) {
+			for (BlockState state : feature.block().getStateDefinition().getPossibleStates()) {
+				state.cache = leafCache;
 			}
 		}
 	}